### 2025-09-18 #9007 @editorialbot
[PRE REVIEW]: GAIO.jl - a concise Julia package for the global analysis of dynamical systems
**Submitting author:** @April-Hannah-Lena (<a href="http://orcid.org/0009-0001-0904-4806">April Herwig</a>)
**Repository:** https://github.com/gaioguys/GAIO.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.1
