### 2025-09-26 #9007 @jbytecode
👋👋👋 Dear @sandreza  and @mbudisic 👋👋👋

**[[PRE REVIEW]: GAIO.jl - a concise Julia package for the global analysis of dynamical systems](https://github.com/openjournals/joss-reviews/issues/9007)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/9007)).

Thank you in advance!
