### 2025-09-29 #9007 @dawbarton
Yes, happy to. Looks interesting.
