### 2025-10-16 #9266 @editorialbot
[REVIEW]: GAIO.jl - A concise Julia package for the global analysis of dynamical systems
**Submitting author:** @April-Hannah-Lena (<a href="http://orcid.org/0009-0001-0904-4806">April Herwig</a>)
**Repository:** https://github.com/gaioguys/GAIO.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
