### 2025-11-04 fc4ad71
add contributor guidelines
