### 2025-11-12 55d9d52
fix boxmeasure plots
