+++ paper/paper.md
@@ -42,7 +42,7 @@ The *maximal invariant set* contained in some set $S$ is
 $$\text{Inv}(S)=\{x\in S\mid f^k(x)\in S\text{ for all }k\in \mathbb{Z}\}.$$
 It follows from the definition that $\text{Inv}(S)$ contains all other invariant sets which are contained in $S$. If $S\subset X$ is forward invariant, then $\text{Inv}(S)=\bigcap_{k\geq 0}f^k(S)$. If $S$ is not forward invariant, the set $A_S:=\bigcap_{k\geq 0}f^k(S)$ is called the *attractor relative to $S$* [@DeHo:97]. This characterization leads to a natural algorithm for approximating $A_S$ by repeatedly tightening a cover by finite collections of subsets of $X$ [@DeHo:97]: 
 
-**Algorithm 2:** 
+**Algorithm 1:** 
     Given a partition $\mathcal{X}$ of $X$ into (essentially) disjoint sets and some covering $\mathcal{A}\subset\mathcal{X}$ of $A_S$, repeat the following two steps until a prescribed diameter of the partition elements is reached:
 
 1. Refine $\mathcal{X}$ into a strictly finer partition $\mathcal{X}'$ (i.e.\ such that $\text{diam}(\mathcal{X}')\leq\theta\cdot\text{diam}(\mathcal{X})$ for some fixed $\theta <1$). Let $\mathcal{A}'$ be the corresponding refinement of the covering $\mathcal{A}$ of $A_S$. 
@@ -63,7 +63,7 @@ For demonstration we compute a long trajectory, shown in the left plot in Figure
 
 ![A trajectory of the four wing system and its (relative) attractor. \label{fig:trajectories}](trajectory_and_attractor.png){ width=90% }
 
-For step 1. in Algorithm 2, we construct a `BoxGrid` `PX` which partitions the domain $X = \left[-5,5\right]^3$ into a $2\times 2\times 2$ grid of boxes. One can think of `PX` as the power set[^power] $\mathcal{P}(\mathcal{X})$ of the partition $\mathcal{X}$: 
+For step 1. in Algorithm 1, we construct a `BoxGrid` `PX` which partitions the domain $X = \left[-5,5\right]^3$ into a $2\times 2\times 2$ grid of boxes. One can think of `PX` as the power set[^power] $\mathcal{P}(\mathcal{X})$ of the partition $\mathcal{X}$: 
 
 ![](code2.png){width=50%}
 
@@ -79,7 +79,7 @@ can then be used for refining the `BoxSet` `A`  by bisecting each box in the $k$
 
 [^power]: The *power set* $\mathcal{P}(S)$ of some set $S$ is the set of all subsets of $S$.
 
-To implement step 2. in Algorithm 2, note that the flow map $f$ induces a map $F:\mathcal{P}(\mathcal{X})\to\mathcal{P}(\mathcal{X})$ via
+To implement step 2. in Algorithm 1, note that the flow map $f$ induces a map $F:\mathcal{P}(\mathcal{X})\to\mathcal{P}(\mathcal{X})$ via
 $$F\left(\left\{\chi\right\}\right)=\left\{\hat\chi\in\mathcal{X}\mid\hat\chi\cap f(\chi)\neq\emptyset\right\}.$$
 
 ![Lifting point maps to maps on boxes. \label{fig:boxcover}](boxcover.png){width=50%}
[…]
