### 2025-12-02 eca0c4a
minor changes in the paper
