### 2025-12-02 #9266 @jbytecode
@April-Hannah-Lena - I've just sent a pull request:

https://github.com/gaioguys/GAIO.jl/pull/129

This PR includes some minor changes and fixes in the paper.

Please review the PR. If you are agree with them, please merge.

Thank you in advance.
