### 2025-12-09 #9266 @jbytecode
@April-Hannah-Lena, @gaioguy - The Zenodo archive currently has the title "gaioguy/GAIO.jl: get DOI from Zenodo". The title of the manuscript as well as the other metadata including author list and their ORCHIDs should match those in the Zenodo archive. Please fix and then ping me. Thank you in advance.
