### 2025-12-09 e41bc60
Update paper.md
