+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: ref.bib
 
 # Summary
 
-We provide an implementation of set-oriented numerical methods [@DeJu:02; @DeFrJu:01] in a Julia package. The package enables the rigorous computation of invariant sets (e.g. chain recurrent sets, attractors and invariant manifolds) in dynamical systems and provides discretizations of the transfer and the Koopman operator, enabling the computation of, e.g., invariant measures, almost invariant, cyclic and coherent sets.  We showcase the features of GAIO.jl by some classical computations of (almost) invariant sets and invariant measures in dynamical systems.
+We provide an implementation of set-oriented numerical methods [@DeJu:02; @DeFrJu:01] in a Julia package. The package enables the rigorous computation of invariant sets (e.g., chain recurrent sets, attractors, invariant manifolds) in dynamical systems and provides discretizations of the transfer and the Koopman operator, enabling the computation of, e.g., invariant measures, almost invariant, cyclic and coherent sets.  We showcase the features of GAIO.jl by some classical computations of (almost) invariant sets and invariant measures in dynamical systems.
 
 # Statement of Need
 
@@ -45,12 +45,12 @@ It follows from the definition that $\text{Inv}(S)$ contains all other invariant
 **Algorithm 1:** 
     Given a partition $\mathcal{X}$ of $X$ into (essentially) disjoint sets and some covering $\mathcal{A}\subset\mathcal{X}$ of $A_S$, repeat the following two steps until a prescribed diameter of the partition elements is reached:
 
-1. Refine $\mathcal{X}$ into a strictly finer partition $\mathcal{X}'$ (i.e.\ such that $\text{diam}(\mathcal{X}')\leq\theta\cdot\text{diam}(\mathcal{X})$ for some fixed $\theta <1$). Let $\mathcal{A}'$ be the corresponding refinement of the covering $\mathcal{A}$ of $A_S$. 
-2. Map the refined covering forward under $f$, i.e. cover [^cover] $f(\vert\mathcal{A}'\vert)$  by elements of $\mathcal{X}'$. Intersect this covering with $\mathcal{A}'$. 
+1. Refine $\mathcal{X}$ into a strictly finer partition $\mathcal{X}'$ (i.e., such that $\text{diam}(\mathcal{X}')\leq\theta\cdot\text{diam}(\mathcal{X})$ for some fixed $\theta <1$). Let $\mathcal{A}'$ be the corresponding refinement of the covering $\mathcal{A}$ of $A_S$. 
+2. Map the refined covering forward under $f$, i.e., cover [^cover] $f(\vert\mathcal{A}'\vert)$  by elements of $\mathcal{X}'$. Intersect this covering with $\mathcal{A}'$. 
 
 [^cover]: $\vert\mathcal{A}\vert =\bigcup_{\chi\in\mathcal{A}}\chi$
 
-In GAIO.jl, partitions resp. coverings are implemented as collections of *boxes*, i.e. multidimensional intervals of the form $\left[\ell_1,u_1\right]\times\ldots\times\left[ \ell_d,u_d\right]$.
+In GAIO.jl, partitions resp. coverings are implemented as collections of *boxes*, i.e., multidimensional intervals of the form $\left[\ell_1,u_1\right]\times\ldots\times\left[ \ell_d,u_d\right]$.
 
 
 ## Example: A four wing attractor
@@ -117,14 +117,14 @@ We compute $F_\sharp$ on the covering constructed above and then compute part of
 
 ![](code8.png){width=49%}
 
[…]
