### 2025-12-09 030460a
Update ref.bib
