+++ paper/ref.bib
@@ -130,9 +130,11 @@
   doi            = {10.1016/C2011-0-06993-4}
 }
 
-@misc{gaiocuda,
+@mastersthesis{gaiocuda,
   author = {A.~Herwig},
-  title  = {{GAIO.jl}: Set-oriented Methods for Approximating Invariant Objects, and their Implementation in {Julia}. \textrm{B.Sc. Thesis, {T}echnical {U}niversity of {M}unich}},
+  title  = {{GAIO.jl}: Set-oriented Methods for Approximating Invariant Objects, and their Implementation in {Julia}.
+  school   = {Technical University of Munich},
+  type   = {B.Sc. Thesis},
   year   = {2022}, 
   url 	 = {github.com/April-Hannah-Lena/schoolwork/blob/main/Bach%20Thesis/main.pdf},
 }
