### 2025-12-11 #9266 @danielskatz
Interesting - I've always done this, and latex/bibtex has never complained, though I agree it's not quite proper.

@gaioguy - Can you change this line to remove the final comma? Then can we can see if it fixes the problem
