### 2025-12-11 890bcb9
Fix formatting of URL in reference entry
