+++ paper/ref.bib
@@ -136,7 +136,7 @@
   school   = {Technical University of Munich},
   type   = {B.Sc. Thesis},
   year   = {2022}, 
-  url 	 = {github.com/April-Hannah-Lena/schoolwork/blob/main/Bach%20Thesis/main.pdf},
+  url 	 = {github.com/April-Hannah-Lena/schoolwork/blob/main/Bach%20Thesis/main.pdf}
 }
 
 @article{dynamicalsystemsjl,
