### 2025-12-11 56e01b0
Fix formatting of title in GAIO.jl reference
