+++ paper/ref.bib
@@ -132,7 +132,7 @@
 
 @mastersthesis{gaiocuda,
   author = {A.~Herwig},
-  title  = {{GAIO.jl}: Set-oriented Methods for Approximating Invariant Objects, and their Implementation in {Julia},
+  title  = {{GAIO.jl}: Set-oriented Methods for Approximating Invariant Objects, and their Implementation in {Julia}},
   school   = {Technical University of Munich},
   type   = {B.Sc. Thesis},
   year   = {2022}, 
