### 2025-12-12 02cabbe
Fix formatting in reference for Rumpf and Wierse
