+++ paper/ref.bib
@@ -321,10 +321,12 @@ doi = {10.1007/978-3-642-56589-2_7},
 
 @article{RuWi92a,
 	author = {M.~Rumpf and A.~Wierse},
-	journal = {Informatik, Forschung und Entwicklung},
+	journal = {Informatik -- Forschung und Entwicklung},
 	pages = {145--151},
-	title = {{GRAPE}, eine objektorientierte {V}isualisierungs-- und {N}u\-me\-rik\-plattform.},
+	title = {{GRAPE}, eine objektorientierte {V}isualisierungs-- und {N}u\-me\-rik\-plattform},
 	volume = {7},
+	number = {3},
+    pages  = {145--151},
 	year = {1992}}
 	
 @Article{harris2020array,
