### 2025-12-12 #9266 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
