### 2025-12-12 #9266 @danielskatz
Congratulations to @April-Hannah-Lena (April Herwig) and @gaioguy and co-author on your JOSS publication!!

JOSS completely depends on volunteers; we could't do this without you
